refactor(database): Fix imports in biorxiv_api.py and factory_design_pattern.py (Albert-2022)
